Add PHAC primer scheme
Add primer scheme from github.com/phac-nml/resende-ncov2020 to repo for
ease of use.

To use these or other schemes, just specify the scheme in the
config.yaml.

Closes #76

Additionally, the 5th field of the primer beds should be the "score" rather than pool
This doesn't seem to be used by ivar but I've fixed it in the included
schemes regardless.