Add freebayes variant/consensus option from @jts
This implements @jts's code
[here](https://github.com/oicr-gsi/ncov2019-artic-nf/compare/master...jts:freebayes#diff-71cfae78f83f02a07ea34cc5b043d519b9162ab7b2d5d62f30a47f4353fac489R135)

This is intended to solve the ivar indel issue in variant calling and
consensus generation.

Freebayes and breseq can each be optionally run via the config file
in addition to the standard ivar workflow.

Currently results from freebayes are not reported in the output report
but can be found in `results_dir/{samplename}/freebayes`