Enable complete run from samples without qc-pass reads
Currently SIGNAL fails for samples with no reads passing QC. This is an
irregular problem that mainly occurs for negative controls when they are
very clean. This is somewhat frustrating to fix and requires manual
intervention. These changes remove two of the failure points.

- Trimgalore rule fails (tool itself exits non-zero) with an empty fastq
  so added a touch to create empty output files if the trimgalore command fails

- Kraken2 rule also fails as kraken2 doesn't produce a report output for
  empty input files (but still does exit 0). Added a touch on the report
  file on successful completion of kraken2 (will not change the content
  of a valid report file and will create a non-existent one).

*NOTE*: workflow will still fail when running quast or comparison script  on
ivar consensus sequence (malformed fasta with just header instead of all
N fasta generated by freebayes). However, ivar is soon to be removed so
only freebayes is run.